benchmarking/llmbenchmarking/pyproject.toml
[project]
name = "llmbenchmarking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<=3.13"
dependencies = [
    "llm",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]

[project.scripts]
llmbenchmarking = "llmbenchmarking.benchmarking:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
llm = { workspace = true }
